move UIserAccount from a repository to a service as to align with services responsibilities.
