+++ ParkBuddy.Application/Interfaces/IUserAccountService.cs
@@ -3,7 +3,7 @@ using ParkBuddy.Contracts.Common;
 
 namespace ParkBuddy.Application.Interfaces;
 
-public interface IUserRepository
+public interface IUserAccountService
 {
     Task<Result<Guid>> RegisterUserAsync(RegisterUserCommand user, CancellationToken cancellationToken);
 }
