+++ ParkBuddy.Infrastructure/Identity/UserAccountService.cs
@@ -5,19 +5,33 @@ using ParkBuddy.Contracts.Common;
 using ParkBuddy.Domain.Entities;
 using ParkBuddy.Infrastructure.Data;
 
-namespace ParkBuddy.Infrastructure.Repositories;
+namespace ParkBuddy.Infrastructure.Identity;
 
-public class UserRepository : IUserRepository
+/// <summary>
+/// Repository for managing user-related operations. 
+/// </summary>
+public class UserAccountService : IUserAccountService
 {
     private readonly UserManager<User> _userManager;
     private readonly ParkBuddyContext _context;
 
-    public UserRepository(ParkBuddyContext context, UserManager<User> userManager)
+    /// <summary>
+    /// Initializes a new instance of the <see cref="UserAccountService"/> class with the specified database context and user manager.
+    /// </summary>
+    /// <param name="context">The database context to be used by the repository.</param>
+    /// <param name="userManager">The user manager to be used by the repository.</param>
+    public UserAccountService(ParkBuddyContext context, UserManager<User> userManager)
     {
         _context = context;
         _userManager = userManager;
     }
 
+    /// <summary>
+    /// Registers a new user with the provided registration details.
+    /// </summary>
+    /// <param name="user">The registration details for the new user.</param>
+    /// <param name="cancellationToken">The cancellation token to monitor for cancellation requests.</param>
+    /// <returns>A result indicating success or failure of the registration operation.</returns>
     public async Task<Result<Guid>> RegisterUserAsync(RegisterUserCommand user, CancellationToken cancellationToken)
     {
         using var transaction = await _context.Database.BeginTransactionAsync();
@@ -36,14 +50,14 @@ public class UserRepository : IUserRepository
         if (!result.Succeeded)
         {
             await transaction.RollbackAsync();
-            return Result<Guid>.Failure("User creation failed");
+            return Result<Guid>.Failure("User creation failed", result.Errors.Select(e => e.Description).ToList());
         }
 
         var roleResult = await _userManager.AddToRoleAsync(newUser, user.Role.ToString());
         if (!roleResult.Succeeded)
         {
             await transaction.RollbackAsync();
-            return Result<Guid>.Failure("Role assignment failed");
+            return Result<Guid>.Failure("Role assignment failed", roleResult.Errors.Select(e => e.Description).ToList());
         }
 
         await transaction.CommitAsync();
